Swap references to hostname for server to identifer in reservation scripts:
This seems like a holdover from the on-computer population script which would default to only using the hostname as an identifier. Since it is now more likely to be a serial number or service tag, this should be changed in documentation. There was already one user who brought this up as a point of confusion given most systems do not have hostnames in the GLPI instance